backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
pymysql==1.1.1
